% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exiftool_locations.R
\name{complete_location}
\alias{complete_location}
\title{Sets iptc and xmp location tags from each other respectively}
\usage{
complete_location(
  paths,
  csv_execute = TRUE,
  csv_path = "~/Pictures/locations.csv",
  delete_original = FALSE
)
}
\arguments{
\item{paths}{List of photos to be modified. Needs a character vector with full file names}

\item{csv_execute}{logical, to determine whether modification should be directly written or returned as tibble}

\item{csv_path}{path and file name of the output csv file}

\item{delete_original}{whether the original copies of the photos should be deleted afterwards or not}
}
\value{
depending on param csv_execute:
\itemize{
\item if TRUE,  writes the tag values as csv file and writes them via exiftool to pictures
\item if FALSE, returns the tags as tibble
}
}
\description{
Sets the tags for city, province and country from IPTC group to XMP group and vice versa.
Goal is to have all available location info identical in both groups. This is necessary because many
programs consider different groups for that information which leads to inconsistencies.
It uses per default the csv option of exifool to handle different values for different files,
which is much faster than a for loop which calls exiftool every time.
Exiftool per default creates copies of the original files (*_original); this behaviour can be modified by parameter.
}
