% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exiftool_convert_to_35mm.R
\name{convert35}
\alias{convert35}
\title{amend crop factor information for focallength bases on a mapping table}
\usage{
convert35(
  paths,
  csv_execute = TRUE,
  csv_path = "~/Pictures/fl.csv",
  delete_original = FALSE
)
}
\arguments{
\item{paths}{List of photos to be modified. Needs a character vector with full file names}

\item{csv_execute}{logical, to determine whether modification should be directly written or returned as tibble}

\item{csv_path}{path and file name of the output csv file}

\item{delete_original}{whether the original copies of the photos should be deleted afterwards or not}
}
\value{
depending on param csv_execute:
\itemize{
\item if TRUE,  writes the tag values as csv file and writes them via exiftool to pictures
\item if FALSE, returns the tags as tibble
}
}
\description{
Writes the tag for equivalent focal length in 35mm format into photo metadata.
Relies on a hardcoded mapping table (CropFactor.csv).
It uses per default the csv option of exifool to handle different values for different files,
which is much faster than a for loop which calls exiftool every time.
Exiftool per default creates copies of the original files (*_original); this behaviour can be modified by parameter.
}
