% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_hours}
\alias{convert_hours}
\title{Time format conversion}
\usage{
convert_hours(hours)
}
\arguments{
\item{hours}{usually the offset of the timezone compared to UTC; a number}
}
\value{
the input parameter in the pre-defined output format
}
\description{
creates the time format used in Exiftool for Offsets (e.g. "+01:00")
}
