% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exiftool_lensinfo.R
\name{create_lensinfo}
\alias{create_lensinfo}
\title{Generates the lens info (for zoom or prime) from the lensmodel}
\usage{
create_lensinfo(lensmodel, as_tags = TRUE)
}
\arguments{
\item{lensmodel}{the lensmodel information from the respective metadata tag (usually from exiftool)}

\item{as_tags}{if TRUE, function returns lensinfo as exiftool arguments, else as raw information}
}
\value{
a tibble with either the raw lensinfo or already prepared as exiftool arguments, depending on param as_tags
}
\description{
Format is: min focal range, max focal range, min aperture @ min range, min aperture @ max range
for primes, V2 and V4 are identical to V1 and V3 respectively
}
