% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cropFactor}
\alias{cropFactor}
\title{Crop Factor data}
\format{
a csv file yielding a data frame
\describe{
\item{model}{the camera model (exif data)}
\item{lensmodel}{the lens model (exif data), only for smartphone cameras relevant and NA for all else}
\item{factor}{the crop factor (focal length times crop factor yields the 35mm equivalent focal length)}
}
}
\source{
self produced based on exif data
}
\usage{
cropFactor
}
\description{
data describes the crop factor of cameras and smartphones
this is needed to compute the 35mm equivalent focal length
}
\keyword{datasets}
