% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exifer-package.R
\docType{package}
\name{exifer-package}
\alias{exifer}
\alias{exifer-package}
\title{exifer: Photo Metadata Improver}
\description{
Updates, harmonizes and synchonizes photo metadata with the help of exiftool Works closely in conjunction with ON1 and PhotoStatistica to achieve coherent metadata across your photo collection.
}
\author{
\strong{Maintainer}: Daniel Baumgartner \email{insilentio@pm.me}

Authors:
\itemize{
  \item Daniel Baumgartner \email{insilentio@pm.me}
}

}
\keyword{internal}
