% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_paths}
\alias{extract_paths}
\title{Read PhotoStatistica exif_export.csv for path creation}
\usage{
extract_paths(input = "~/Pictures/Exif-export.csv")
}
\arguments{
\item{input}{path to the exif-export csv file}
}
\value{
the calculated paths as a tibble
}
\description{
The function extracts the different info from the exif_export file from PhotoStatistica
and calculates the indidividual file paths. They are then written to a file which can be
used with exiftool with the "-@" option
}
