% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{handle_return}
\alias{handle_return}
\title{Handles the outcome of the various exiftool functions}
\usage{
handle_return(
  df,
  csv_execute,
  paths,
  csv_path,
  delete_original,
  with_sep = NULL
)
}
\arguments{
\item{df}{tibble or dataframe with the file and metadata information. Usually created within one
of the exiftool_ file functions.}

\item{csv_execute}{logical, to determine whether modification should be directly written or returned as tibble}

\item{paths}{List of photos to be modified. Needs a character vector with full file names}

\item{csv_path}{path and file name of the output csv file}

\item{delete_original}{whether the original copies of the photos should be deleted afterwards or not}

\item{with_sep}{for list type tags like keywords, you need to indicate which separator is used}
}
\value{
depending on param csv_execute:
\itemize{
\item if TRUE,  writes the tag values as csv file and writes them via exiftool to pictures
\item if FALSE, returns the tags as tibble
}
}
\description{
depending on param csv_execute, either writes a csv_file with per photo specific metadata
and runs exiftool on top of that to update the photos or returns the respective information as a tibble.
}
