% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exiftool_lensinfo.R
\name{harmonize_lensinfo}
\alias{harmonize_lensinfo}
\title{Write lens information}
\usage{
harmonize_lensinfo(
  paths,
  csv_execute = TRUE,
  csv_path = "~/Pictures/lensinfo.csv",
  delete_original = FALSE,
  lensinfo_only = FALSE
)
}
\arguments{
\item{paths}{List of photos to be modified. Needs a character vector with full file names}

\item{csv_execute}{logical, to determine whether modification should be directly written or returned as tibble}

\item{csv_path}{path and file name of the output csv file}

\item{delete_original}{whether the original copies of the photos should be deleted afterwards or not}

\item{lensinfo_only}{if TRUE, only the tag "lensinfo" is changed (this is often omitted by many programs)}
}
\value{
depending on param csv_execute:
\itemize{
\item if TRUE,  writes the tag values as csv file and writes them via exiftool to pictures
\item if FALSE, returns the tags as tibble
}
}
\description{
Writes various lens information tags obtained from a mapping table (LensMapping.csv) and existing metadata
into photo metadata. LensInfo is created dynamically and can be written separately by the according parameter
(this is often necessary because programs omit this information which is relevant for prime/zoom differentiation).
It uses per default the csv option of exifool to handle different values for different files,
which is much faster than a for loop which calls exiftool every time
}
