% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exiftool_time.R
\name{harmonize_time}
\alias{harmonize_time}
\title{sets datetimeoriginal to createdate}
\usage{
harmonize_time(
  paths,
  csv_execute = TRUE,
  csv_path = "~/Pictures/time.csv",
  delete_original = FALSE
)
}
\arguments{
\item{paths}{List of photos to be modified. Needs a character vector with full file names}

\item{csv_execute}{logical, to determine whether modification should be directly written or returned as tibble}

\item{csv_path}{path and file name of the output csv file}

\item{delete_original}{whether the original copies of the photos should be deleted afterwards or not}
}
\value{
depending on param csv_execute:
\itemize{
\item if TRUE,  writes the tag values as csv file and writes them via exiftool to pictures
\item if FALSE, returns the tags as tibble
}
}
\description{
Sets the datetimeoriginal tag to the same value as createdate to obtain more coherent time of creation.
Especially important for mixed camera scenarioes.
Does not solve all time related issues yet, needs more work and clarification with regard to ON1.
}
