% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lensMapping}
\alias{lensMapping}
\title{Lens mapping data}
\format{
a csv file yielding a data frame
\describe{
\item{model_old}{the old/original camera model name (exif data)}
\item{model_new}{the new, harmonized camera model name}
}
}
\source{
self produced based on exif data
}
\usage{
lensMapping
}
\description{
data describes the the mapping from unharmonized camera names to harmonized names
}
\keyword{datasets}
