% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{output_paths}
\alias{output_paths}
\title{Write complete path names into csv file}
\usage{
output_paths(paths, output = "~/Pictures/paths.txt")
}
\arguments{
\item{paths}{a tibble of path information obtained by extract_paths()}

\item{output}{path to the desired output csv file}
}
\value{
Invisibly writes a csv file into output
}
\description{
Write complete path names into csv file
}
