% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_export.R
\name{prepare_export}
\alias{prepare_export}
\title{Prepare originals for export}
\usage{
prepare_export(
  imp_path = "/Volumes/NoBackup/Bilder/Import/2025/",
  level_below = "1",
  offset
)
}
\arguments{
\item{imp_path}{current import path to work on. The number of subordinate directory levels is indicated
by the parameter 'level_below'.}

\item{level_below}{indicates to which directory level relative to imp_path it is looking for xmp files and
therefore may need to be changed (e.g. on year level -> 2, on album level -> 3)}

\item{offset}{the number of hours of offset of the timezone the photos were taken in compared to UTC.
1 for MEZ, 2 for MESZ}
}
\value{
nothing
}
\description{
Performs necessary steps on metadata side for new pictures
in Import folder after Development work in ON1 and before Export
\enumerate{
\item copy xmp metadata information onto original file
\item remove unnecessary side files
\item flatten hierarchical keywords for writing into keywords and subject tag
\item amend lens information for Nikon in order to have consistent metadata
\item amend lens information for Apple in order to have prime/zoom lens info consistent
\item ensures completeness of location tags information
\item harmonize createtime and datetimeoriginal information
}
}
